Deleted Email and Birthday input fields
This example is not concerned with the actual content of the portlet, just
to show how the functionality of the portlet changes when the script
element comes first. So, the contents are changed to be the same as the
past two exercises (9 and 10), so the focus is on the difference in the
script position.